refactor functions in to helpers
